Fix #23 - Decouple events and boundaries in .removeAll()
For increased clearity in API it make sense to be consistent
in naming. If .remove() only remove boundary without touching
event listeners, .removeAll() should do the same thing.
Also with previous approach there were no way to flush all boundaries
without also flushing the events. If you work with things such as
redux-saga it makes a lot of sense to decouple theese things because
the events are probably best initialized on app startup but boundaries
are probably initialized/replaced on any API-response.